pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sqlalchemy[asyncio]>=2.0.48",
    "temporalio>=1.23.0",
    "uvicorn[standard]>=0.41.0",
    "pyjwt[crypto]>=2.11.0",
    "passlib[argon2]>=1.7.4",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
